Add user file
What: This file will contain the functions to store, load and update the users
Why: Functionality and readability

JOHAN
