--- .idea/libraries/jfoenix_9_0_8.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="jfoenix-9.0.8">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/librerias/jfoenix-9.0.8.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
